# API_Factory/Sentiment_Alpha_v1/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arsenal Decision Engine \u2014 Decision Intelligence Layer",
        "description": "Predict freight volatility before it hits your budget. Analyzes global news to generate a Market Psychology Index and resolve conflicting signals via an AI-Conflict Resolution Layer.",
        "version": "4.0.0"
    },
    "paths": {
        "/analyze": {
            "post": {
                "summary": "Analyze Market Mood",
                "description": "Standard Financial Fear & Greed Index.",
                "operationId": "analyze_market_mood_analyze_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Analyze Market Mood Analyze Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/maritime": {
            "post": {
                "summary": "Analyze Maritime Risk",
                "description": "Maritime freight risk analysis using quantitative psychology principles.\nPredicts price surges based on global routing tensions.\nCan filter by zone (e.g., 'transpacific').",
                "operationId": "analyze_maritime_risk_analyze_maritime_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaritimeInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Analyze Maritime Risk Analyze Maritime Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/compare": {
            "post": {
                "summary": "Compare Corridors Risk",
                "description": "Compares the Transpacific and Suez Europe corridors to identify arbitrage opportunities.",
                "operationId": "compare_corridors_risk_analyze_compare_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeridianOutput"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/arbitrage": {
            "post": {
                "tags": [
                    "Decision Engine"
                ],
                "summary": "Evaluate Arbitrage Opportunity",
                "description": "**Quantitative Arbitrage Decision Engine** \u2014 GEMINI.md \u00a74 + Quantitative Decision Matrix.\n\nEvaluates a DeFi arbitrage opportunity through 6 sequential filters:\n- F1: Risk/Reward asymmetry \u2265 3:1\n- F2: Breakout confirmation (4H/Daily close above resistance)\n- F3: Volume anomaly V_24h > SMA(V,20) + 2\u03c3\n- F4: ATR positive slope over 3 consecutive periods\n- F5: Funding rate < 0.05% per 8h (long squeeze guard)\n- F6: Stop-loss node mandatory (capital preservation rule)\n\nThen computes R_net = Y_yield + I_incentives \u2212 IL \u2212 G_gas \u2212 S_MEV \u2212 F_funding.\nProposes EXECUTE only if R_net > flash_loan_opportunity_cost AND all 6 filters pass.",
                "operationId": "evaluate_arbitrage_opportunity_analyze_arbitrage_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArbitrageInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Evaluate Arbitrage Opportunity Analyze Arbitrage Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/mcp/v1/tools": {
            "get": {
                "tags": [
                    "MCP Bridge"
                ],
                "summary": "List Mcp Tools",
                "description": "Endpoint standard du protocole MCP permettant aux agents autonomes (Claude, etc.)\nde d\u00e9couvrir les capacit\u00e9s et les sch\u00e9mas de Meridian Alpha en t\u00e2che de fond.",
                "operationId": "list_mcp_tools_mcp_v1_tools_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mcp/v1/tools/execute": {
            "post": {
                "tags": [
                    "MCP Bridge"
                ],
                "summary": "Execute Mcp Tool",
                "description": "Point d'ex\u00e9cution automatique pour les agents IA autonomes.",
                "operationId": "execute_mcp_tool_mcp_v1_tools_execute_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MCPExecuteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "MCPExecuteRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "arguments": {
                        "type": "object",
                        "title": "Arguments"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "arguments"
                ],
                "title": "MCPExecuteRequest"
            },
            "MaritimeInput": {
                "properties": {
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "zone": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Zone"
                    }
                },
                "type": "object",
                "required": [
                    "text"
                ],
                "title": "MaritimeInput"
            },
            "MeridianOutput": {
                "properties": {
                    "most_volatile_corridor": {
                        "type": "string",
                        "title": "Most Volatile Corridor"
                    },
                    "transpacific": {
                        "$ref": "#/components/schemas/RouteAnalysis"
                    },
                    "suez_europe": {
                        "$ref": "#/components/schemas/RouteAnalysis"
                    },
                    "arbitration_required": {
                        "type": "boolean",
                        "title": "Arbitration Required"
                    },
                    "arbitration_reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Arbitration Reason"
                    }
                },
                "type": "object",
                "required": [
                    "most_volatile_corridor",
                    "transpacific",
                    "suez_europe",
                    "arbitration_required"
                ],
                "title": "MeridianOutput"
            },
            "NewsInput": {
                "properties": {
                    "text": {
                        "type": "string",
                        "title": "Text"
                    }
                },
                "type": "object",
                "required": [
                    "text"
                ],
                "title": "NewsInput"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}